task correction main.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 
 /**
- * linear_search - Searches for a value in an array of integers using linear search
- * @array: Pointer to the first element of the array to search in
- * @size: Number of elements in the array
- * @value: Value to search for
+ * linear_search - Searches for a value in an array of integers using Linear search algorithm.
+ * @array: Pointer to the first element of the array to search in.
+ * @size: Number of elements in the array.
+ * @value: Value to search for.
  *
- * Return: The first index where value is located, or -1 if not found or array is NULL
+ * Return: The first index where value is located, or -1 if value is not present or array is NULL.
  */
 int linear_search(int *array, size_t size, int value)
 {
@@ -20,8 +20,7 @@ int linear_search(int *array, size_t size, int value)
     {
         printf("Value checked array[%lu] = [%d]\n", i, array[i]);
         if (array[i] == value)
-            return ((int)i);
+            return (i);
     }
-
     return (-1);
 }
